New LLM provider and models config file (#61)
Models can now be selected from a dropdown below the input area. Most CLI args to configure providers have been replaced by the config file mechanism. You can provide a model (by name) on the command line. But otherwise, the model is stored with the session and will be used when continuing a session.